Move base64 code to its own file.
I haven't tested the build with bazel but the normal build process works and
tests are passing.
