+++ BUILD.bazel
@@ -124,6 +124,7 @@ cc_library(
         "libyara/ahocorasick.c",
         "libyara/arena.c",
         "libyara/atoms.c",
+        "libyara/base64.c",
         "libyara/bitmask.c",
         "libyara/compiler.c",
         "libyara/crypto.h",
@@ -141,6 +142,7 @@ cc_library(
         "libyara/include/yara/ahocorasick.h",
         "libyara/include/yara/arena.h",
         "libyara/include/yara/atoms.h",
+        "libyara/include/yara/base64.h",
         "libyara/include/yara/bitmask.h",
         "libyara/include/yara/compiler.h",
         "libyara/include/yara/dex.h",
