+++ libyara/Makefile.am
@@ -95,6 +95,7 @@ yarainclude_HEADERS = \
 	include/yara/ahocorasick.h \
 	include/yara/arena.h \
 	include/yara/atoms.h \
+	include/yara/base64.h \
 	include/yara/bitmask.h \
 	include/yara/compiler.h \
 	include/yara/error.h \
@@ -158,6 +159,7 @@ libyara_la_SOURCES = \
 	ahocorasick.c \
 	arena.c \
 	atoms.c \
+	base64.c \
 	bitmask.c \
 	compiler.c \
 	endian.c \
