+++ libyara/include/yara/re.h
@@ -105,12 +105,6 @@ typedef int RE_MATCH_CALLBACK_FUNC(
     int flags,
     void* args);
 
-int yr_re_ast_from_base64(
-    SIZED_STRING* str,
-    YR_MODIFIER modifier,
-    RE_AST** re_ast,
-    RE_ERROR* re_error);
-
 int yr_re_ast_create(
     RE_AST** re_ast);
